moved most mail creation functionality to newsletter.js
